ref(server): Rename metrics module to statsd
