+++ .github/workflows/deploy.yml
@@ -124,7 +124,7 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: run
-          args: -p document-metrics -- -o relay_metrics.json relay-server/src/metrics.rs
+          args: -p document-metrics -- -o relay_metrics.json relay-server/src/statsd.rs
 
       - name: Deploy
         if: github.ref == 'refs/heads/master'
