+++ relay-server/src/utils/kafka.rs
@@ -4,7 +4,7 @@ use rdkafka::ClientContext;
 use relay_log::LogError;
 use relay_statsd::metric;
 
-use crate::metrics::RelayCounters;
+use crate::statsd::RelayCounters;
 
 /// Kafka producer context that logs producer errors
 pub struct CaptureErrorContext;
